Finished implementing the Fragment for Country API Still failed Observer
