+++ app/src/main/java/com/example/android/myapplication/CountryObserver.java
@@ -1,7 +1,6 @@
 package com.example.android.myapplication;
 
 import android.app.Activity;
-import android.widget.TextView;
 
 import java.util.Observable;
 import java.util.Observer;
@@ -18,20 +17,45 @@ public class CountryObserver extends Activity implements Observer {
     public void update(Observable o, Object arg) {
 
         countryModel = (CountryModel) o;
-
-        TextView nameTextView = (TextView) findViewById(R.id.name);
-        nameTextView.setText(countryModel.getName());
-
-        TextView currencyTextView = (TextView) findViewById(R.id.currency);
-        currencyTextView.setText(countryModel.getCurrency());
-
-        TextView populationTextView = (TextView) findViewById(R.id.population);
-        populationTextView.setText(Integer.toString(countryModel.getPopulation()));
-
-        TextView languageTextView = (TextView) findViewById(R.id.language);
-        languageTextView.setText(countryModel.getLanguage());
-
-        TextView capitalTextView = (TextView) findViewById(R.id.capital);
-        capitalTextView.setText(countryModel.getCapital());
+//        final TextView nameTextView = (TextView) findViewById(R.id.name);
+//
+//        nameTextView.post(new Runnable() {
+//            @Override
+//            public void run() {
+//                nameTextView.setText(countryModel.getName());
+//            }
+//        });
+//
+//        final TextView currencyTextView = (TextView) findViewById(R.id.currency);
+//        currencyTextView.post(new Runnable() {
+//            @Override
+//            public void run() {
+//                currencyTextView.setText(countryModel.getCurrency());
+//            }
+//        });
+//
+//        final TextView populationTextView = (TextView) findViewById(R.id.population);
+//        populationTextView.post(new Runnable() {
+//            @Override
+//            public void run() {
+//                populationTextView.setText(Integer.toString(countryModel.getPopulation()));
+//            }
+//        });
+//
+//        final TextView languageTextView = (TextView) findViewById(R.id.language);
+//        languageTextView.post(new Runnable() {
+//            @Override
+//            public void run() {
+//                languageTextView.setText(countryModel.getLanguage());
+//            }
+//        });
+//
+//        final TextView capitalTextView = (TextView) findViewById(R.id.capital);
+//        capitalTextView.post(new Runnable() {
+//            @Override
+//            public void run() {
+//                capitalTextView.setText(countryModel.getCapital());
+//            }
+//        });
     }
 }
